+++ models/src/main/java/io/sphere/sdk/products/commands/updateactions/SetSku.java
@@ -6,10 +6,10 @@ import io.sphere.sdk.products.Product;
 import java.util.Optional;
 
 public class SetSku extends UpdateAction<Product> {
-    private final long variantId;
+    private final int variantId;
     private final Optional<String> sku;
 
-    private SetSku(final long variantId, final Optional<String> sku) {
+    private SetSku(final int variantId, final Optional<String> sku) {
         super("setSKU");
         this.variantId = variantId;
         this.sku = sku;
@@ -23,11 +23,11 @@ public class SetSku extends UpdateAction<Product> {
         return sku;
     }
 
-    public static SetSku of(final long variantId, final Optional<String> sku) {
+    public static SetSku of(final int variantId, final Optional<String> sku) {
         return new SetSku(variantId, sku);
     }
 
-    public static SetSku of(final long variantId, final String sku) {
+    public static SetSku of(final int variantId, final String sku) {
         return of(variantId, Optional.of(sku));
     }
 }
